Stardew Valley: Move BaseTest out of `__init__.py` to comply with future conventions (#4991)
* move everything out of init; fix from imports and some typing errors

* why is there a change in multiserver

* fix some relative shits
