+++ worlds/stardew_valley/test/TestFishsanity.py
@@ -1,8 +1,8 @@
 import unittest
 from typing import ClassVar, Set
 
-from . import SVTestBase
 from .assertion import WorldAssertMixin
+from .bases import SVTestBase
 from ..content.feature import fishsanity
 from ..mods.mod_data import ModNames
 from ..options import Fishsanity, ExcludeGingerIsland, Mods, SpecialOrderLocations, Goal, QuestLocations
