+++ worlds/stardew_valley/test/TestOptionFlags.py
@@ -1,4 +1,4 @@
-from . import SVTestBase
+from .bases import SVTestBase
 from .. import BuildingProgression
 from ..options import ToolProgression
