+++ worlds/stardew_valley/test/long/TestModsLong.py
@@ -4,8 +4,8 @@ from typing import ClassVar
 
 from BaseClasses import get_seed
 from test.param import classvar_matrix
-from .. import SVTestCase, solo_multiworld, skip_long_tests
 from ..assertion import WorldAssertMixin, ModAssertMixin
+from ..bases import skip_long_tests, SVTestCase, solo_multiworld
 from ..options.option_names import all_option_choices
 from ... import options
 from ...mods.mod_data import ModNames
