+++ worlds/stardew_valley/test/mods/TestSVE.py
@@ -1,4 +1,4 @@
-from .. import SVTestBase
+from ..bases import SVTestBase
 from ... import options
 from ...mods.mod_data import ModNames
 from ...strings.ap_names.mods.mod_items import SVEQuestItem
