+++ worlds/stardew_valley/test/script/benchmark_locations.py
@@ -15,8 +15,9 @@ import typing
 
 from BaseClasses import CollectionState, Location
 from Utils import init_logging
-from worlds.stardew_valley.stardew_rule.rule_explain import explain
-from ... import test
+from ..bases import setup_solo_multiworld
+from ..options import presets
+from ...stardew_rule.rule_explain import explain
 
 
 def run_locations_benchmark():
@@ -56,12 +57,12 @@ def run_locations_benchmark():
                 parser.add_argument('--state', help="Define the state in which the location will be benchmarked.", type=str, default=None)
                 args = parser.parse_args()
                 options_set = args.options
-                options = getattr(test, options_set)()
+                options = getattr(presets, options_set)()
                 seed = args.seed
                 location = args.location
                 state = args.state
 
-                multiworld = test.setup_solo_multiworld(options, seed)
+                multiworld = setup_solo_multiworld(options, seed)
                 gc.collect()
 
                 if location:
