+++ worlds/stardew_valley/test/stability/TestUniversalTracker.py
@@ -1,9 +1,11 @@
 import unittest
 from unittest.mock import Mock
 
-from .. import SVTestBase, fill_namespace_with_default, skip_long_tests
+from ..bases import skip_long_tests, SVTestBase
 from ..options.presets import allsanity_mods_6_x_x
-from ... import STARDEW_VALLEY, FarmType, BundleRandomization, EntranceRandomization
+from ..options.utils import fill_namespace_with_default
+from ... import STARDEW_VALLEY
+from ...options import FarmType, BundleRandomization, EntranceRandomization
 
 
 @unittest.skipIf(skip_long_tests(), "Long tests disabled")
